remove old methods to parse Nakamura's corresp and match files
