+++ partitura/io/importnakamura.py
@@ -1,4 +1,3 @@
-#!/usr/bin/env python
 # -*- coding: utf-8 -*-
 """
 This module contains methods for parsing score-to-performance alignments
@@ -11,119 +10,80 @@ References
        Alignment"
 """
 import logging
+import re
 import numpy as np
 
-from partitura.performance import PerformedPart
+from partitura.utils import note_name_to_midi_pitch
+from partitura.utils.music import SIGN_TO_ALTER
 
 LOGGER = logging.getLogger(__name__)
 
-
-class MatchError(Exception):
-    pass
+NAME_PATT = re.compile(r"([A-G]{1})([xb\#]*)(\d+)")
 
 
-class NakamuraMatchLine(object):
-    field_names = [
-        "number",
-        "onset",
-        "offset",
-        "notename",
-        "velocityonset",
-        "velocityoffset",
-        "channel",
-        "matchstatus",
-        "scoretime",
-        "snoteID",
-        "errorindex",
-        "skipindex",
-    ]
+def load_nakamuracorresp(fn):
+    """Load a corresp file as returned by Nakamura et al.'s MIDI to MIDI alignment.
 
-    def __init__(
-        self,
-        number,
-        onset,
-        offset,
-        notename,
-        velocityonset,
-        velocityoffset,
-        channel,
-        matchstatus,
-        scoretime,
-        snoteID,
-        errorindex,
-        skipindex,
-    ):
-        self.number = int(number)
-        self.onset = float(onset)
-        self.offset = float(offset)
-        self.velocityonset = int(velocityonset)
-        self.notename = notename
-        self.scoretime = float(scoretime)
-        self.snoteID = snoteID
-
-    @classmethod
-    def from_line(cls, matchline, pos=0):
-        line_split = matchline.split("\t")
-        if len(line_split) != 12:
-            return None
-        else:
-            kwargs = dict(zip(cls.field_names, line_split))
-            match_line = cls(**kwargs)
-            return match_line
-
-
-def parse_nakamuramatchline(line):
-    """
-    Return objects representing the line as line or comment
+    Fields of the file format as specified in [8]_:
+    (ID) (onset time) (spelled pitch) (integer pitch) (onset velocity)
 
     Parameters
     ----------
-    line : str
-        Line of the match file
+    fn : str
+        The nakamura match.txt-file
 
     Returns
     -------
-    nakamuramatchline : subclass of `NakamuraMatchLine`
-       Object representing the line.
+    align : structured array
+        structured array of performed notes
+    ref : structured array
+        structured array of score notes
+    alignment : list
+        The score--performance alignment, a list of dictionaries
     """
+    note_array_dtype = [("onset_sec", "f4"), ("pitch", "i4"), ("id", "U256")]
+    dtype = [
+        ("alignID", "U256"),
+        ("alignOntime", "f"),
+        ("alignSitch", "U256"),
+        ("alignPitch", "i"),
+        ("alignOnvel", "i"),
+        ("refID", "U256"),
+        ("refOntime", "f"),
+        ("refSitch", "U256"),
+        ("refPitch", "i"),
+        ("refOnvel", "i"),
+    ]
+    result = np.loadtxt(fn, dtype=dtype, comments="//")
 
-    from_NakamuraMatchLine_methods = [NakamuraMatchLine.from_line]
-    nakamuramatchline = False
-    for from_NakamuraMatchLine in from_NakamuraMatchLine_methods:
-        try:
-            nakamuramatchline = from_NakamuraMatchLine(line)
-            break
-        except MatchError:
-            continue
+    align_valid = result["alignID"] != "*"
+    n_align = sum(align_valid)
+    align = np.empty((n_align,), dtype=note_array_dtype)
+    align[:] = result[["alignOntime", "alignPitch", "alignID"]][align_valid]
 
-    return nakamuramatchline
+    ref_valid = result["refID"] != "*"
+    n_ref = sum(ref_valid)
+    ref = np.empty((n_ref,), dtype=note_array_dtype)
+    ref[:] = result[["refOntime", "refPitch", "refID"]][ref_valid]
 
+    alignment = []
+    for alignID, refID in result[["alignID", "refID"]]:
+        if alignID == "*":
+            alnote = dict(label="deletion",
+                          score_id=refID)
+        elif refID == "*":
+            alnote = dict(label="insertion",
+                          performance_id=alignID)
+        else:
+            alnote = dict(label="match",
+                          score_id=refID,
+                          performance_id=alignID)
+        alignment.append(alnote)
 
-class NakamuraMatchFile(object):
-    """
-    Class for representing Nakamura et al.'s match.txt Files
-    """
-
-    def __init__(self, filename):
-        self.name = filename
-
-        with open(filename) as f:
-            self.lines = np.array(
-                [parse_nakamuramatchline(line) for line in f.read().splitlines()]
-            )
-
-    def iter_notes(self):
-        """
-        Iterate over all performed notes
-        """
-        for x in self.lines:
-            if x is None:
-                continue
-            else:
-                yield x
+    return align, ref, alignment
 
 
-def load_nakamuramatch(fn, pedal_threshold=64, first_note_at_zero=False):
+def load_nakamuramatch(fn):
     """Load a match file as returned by Nakamura et al.'s MIDI to musicxml alignment
 
     Fields of the file format as specified in [8]_:
@@ -134,17 +94,13 @@ def load_nakamuramatch(fn, pedal_threshold=64, first_note_at_zero=False):
     ----------
     fn : str
         The nakamura match.txt-file
-    pedal_threshold : int, optional
-        Threshold for adjusting sound off of the performed notes using
-        pedal information. Defaults to 64.
-    first_note_at_zero : bool, optional
-        When True the note_on and note_off times in the performance
-        are shifted to make the first note_on time equal zero.
 
     Returns
     -------
-    ppart : list
-        The performed part, a list of dictionaries
+    align : structured array
+        structured array of performed notes
+    ref : structured array
+        structured array of score notes
     alignment : list
         The score--performance alignment, a list of dictionaries
 
@@ -153,349 +109,160 @@ def load_nakamuramatch(fn, pedal_threshold=64, first_note_at_zero=False):
     .. [8] https://midialignment.github.io/MANUAL.pdf
 
     """
-    # Parse Matchfile
-    mf = NakamuraMatchFile(fn)
-
-    # Generate PerformedPart
-    ppart = performed_part_from_nakamuramatch(mf, pedal_threshold, first_note_at_zero)
-
-    # Alignment
-    alignment = alignment_from_nakamuramatch(mf)
-
-    return mf, ppart, alignment
-
-
-def alignment_from_nakamuramatch(mf):
-    result = []
-    for line in mf.iter_notes():
-        result.append(dict(label="match",
-                           score_id=line.snoteID,
-                           performance_id=line.number))
-    return result
-
-
-def performed_part_from_nakamuramatch(mf, pedal_threshold=64, first_note_at_zero=False):
-    """Make PerformedPart from performance info in a match.txt file
-
-    Parameters
-    ----------
-    mf : nakamuramatchfile
-        A MatchFile instance
-    pedal_threshold : int, optional
-        Threshold for adjusting sound off of the performed notes using
-        pedal information. Defaults to 64.
-    first_note_at_zero : bool, optional
-        When True the note_on and note_off times in the performance
-        are shifted to make the first note_on time equal zero.
-
-    Returns
-    -------
-    ppart : PerformedPart
-        A performed part
-
-    """
-    notes = []
-
-    first_note = next(mf.iter_notes(), None)
-    if first_note and first_note_at_zero:
-        offset = first_note.onset
-    else:
-        offset = 0
-
-    for note in mf.iter_notes():
-        notes.append(
-            dict(
-                id=note.number,
-                midi_pitch=note_name_to_midi_pitch(note.notename),
-                note_on=note.onset - offset,
-                note_off=note.offset - offset,
-                sound_off=note.offset - offset,
-                velocity=note.velocityonset,
-            )
-        )
-
-    ppart = PerformedPart(
-        id="P1",
-        part_name="unknown_part_1",
-        notes=notes,
-        sustain_pedal_threshold=pedal_threshold,
-    )
-    return ppart
-
-
-class NakamuraCorrespLine(object):
-    field_names = [
-        "alignID",
-        "alignOntime",
-        "alignSitch",
-        "alignPitch",
-        "alignOnvel",
-        "refID",
-        "refOntime",
-        "refSitch",
-        "refPitch",
-        "refOnvel",
+    perf_dtype = [("onset_sec", "f4"),
+                  ("duration_sec", "f4"),
+                  ("pitch", "i4"),
+                  ("velocity", "i4"),
+                  ("channel", "i4"),
+                  ("id", "U256"),
+                  ]
+    score_dtype = [("onset_div", "i4"),
+                   ("pitch", "i4"),
+                   ("step", "U256"),
+                   ("alter", "i4"),
+                   ("octave", "i4"),
+                   ("id", "U256")
+                   ]
+    dtype = [
+        ("alignID", "U256"),
+        ("alignOntime", "f"),
+        ("alignOfftime", "f"),
+        ("alignSitch", "U256"),
+        ("alignOnvel", "i"),
+        ("alignOffvel", "i"),
+        ("alignChannel", "i"),
+        ("matchstatus", "i"),
+        ("refOntime", "f"),
+        ("refID", "U256"),
+        ("errorindex", "i"),
+        ("skipindex", "U256"),
     ]
+    dtype_missing = [("refOntime", "f"),
+                     ("refID", "U256")]
+    pattern = r"//Missing\s(\d+)\t(.+)"
+    # load alignment notes
+    result = np.loadtxt(fn, dtype=dtype, comments="//")
+    # load missing notes
+    missing = np.fromregex(fn, pattern, dtype=dtype_missing)
 
-    out_pattern = (
-        "{alignID}\t{alignOntime}\t{alignSitch}"
-        "\t{alignPitch}\t{alignOnvel}"
-        "\t{refID}\t{refOntime}"
-        "\t{refSitch}\t{refPitch}\t{refOnvel}"
-    )
+    midi_pitch = np.array(
+        [note_name_to_midi_pitch(n.replace('#', r'\#'))
+         for n in result["alignSitch"]])
 
-    def __init__(
-        self,
-        alignID,
-        alignOntime,
-        alignSitch,
-        alignPitch,
-        alignOnvel,
-        refID,
-        refOntime,
-        refSitch,
-        refPitch,
-        refOnvel,
-    ):
+    align_valid = result["alignID"] != "*"
+    n_align = sum(align_valid)
+    align = np.empty((n_align,), dtype=perf_dtype)
+    align["id"] = result["alignID"][align_valid]
+    align["onset_sec"] = result["alignOntime"]
+    align["duration_sec"] = (result["alignOfftime"][align_valid] -
+                             result["alignOntime"][align_valid])
+    align["pitch"] = midi_pitch[align_valid]
+    align["velocity"] = result["alignOnvel"][align_valid]
+    align["channel"] = result["alignChannel"][align_valid]
 
-        self.id0 = str(alignID)
-        self.onset0 = float(alignOntime)
-        self.pitch0 = int(alignPitch)
-        self.alignSitch = str(alignSitch)
-        self.alignOnvel = int(alignOnvel)
+    ref_valid = result["refID"] != "*"
+    n_valid = sum(ref_valid)
+    n_ref = n_valid + len(missing)
+    ref = np.empty((n_ref,), dtype=score_dtype)
 
-        self.id1 = str(refID)
-        self.onset1 = float(refOntime)
-        self.pitch1 = int(refPitch)
-        self.refSitch = str(refSitch)
-        self.refOnvel = int(refOnvel)
+    ref["id"][:n_valid] = result["refID"][ref_valid]
+    ref["id"][n_valid:] = missing["refID"]
+    ref["onset_div"][:n_valid] = result["refOntime"][ref_valid]
+    ref["onset_div"][n_valid:] = missing["refOntime"]
+    ref["pitch"][:n_valid] = midi_pitch[ref_valid]
+    ref["pitch"][n_valid:] = -1
+    pitch_spelling = [NAME_PATT.search(nn).groups()
+                      for nn in result["alignSitch"]]
 
-    @classmethod
-    def from_line(cls, correspline, pos=0):
-        line_split = correspline.split("\t")
-        del line_split[-1]
+    pitch_spelling = np.array(
+        [(ps[0], SIGN_TO_ALTER[ps[1] if ps[1] != "" else "n"], int(ps[2]))
+         for ps in pitch_spelling])
+    # add pitch spelling information
+    ref["step"][:n_valid] = pitch_spelling[ref_valid][:, 0]
+    ref["alter"][:n_valid] = pitch_spelling[ref_valid][:, 1]
+    ref["octave"][:n_valid] = pitch_spelling[ref_valid][:, 2]
+    # * indicates that is a missing pitch
+    ref["step"][n_valid:] = '*'
 
-        if len(line_split) != 10:
-            return None
+    alignment = []
+    for alignID, refID in result[["alignID", "refID"]]:
+        if alignID == "*":
+            alnote = dict(label="deletion",
+                          score_id=refID)
+        elif refID == "*":
+            alnote = dict(label="insertion",
+                          performance_id=alignID)
         else:
-            kwargs = dict(zip(cls.field_names, line_split))
-            corresp_line = cls(**kwargs)
-            return corresp_line
+            alnote = dict(label="match",
+                          score_id=refID,
+                          performance_id=alignID)
+        alignment.append(alnote)
 
-    @property
-    def corresp_line(self):
-        self.out_pattern.format(
-            alignID=self.id0,
-            alignOntime=self.onset0,
-            alignPitch=self.pitch0,
-            alignSitch=self.alignSitch,
-            alignOnvel=self.alignOnvel,
-            refID=self.id1,
-            refOntime=self.onset1,
-            refPitch=self.pitch1,
-            refSitch=self.refSitch,
-            refOnvel=self.refOnvel,
-        )
+    for refID in missing["refID"]:
+        alignment.append(dict(label="deletion",
+                              score_id=refID))
+
+    return align, ref, alignment
 
 
-def parse_nakamuracorrespline(line):
-    """
-    Return objects representing the line as line or comment
-
-    Parameters
-    ----------
-    line : str
-        Line of the match file
-
-    Returns
-    -------
-    nakamuracorrespline : `NakamuraCorrespLine`
-       Object representing the line.
-    """
-
-    from_NakamuraCorrespLine_methods = [NakamuraCorrespLine.from_line]
-    nakamuracorrespline = None
-    for from_NakamuraCorrespLine in from_NakamuraCorrespLine_methods:
-        try:
-            nakamuracorrespline = from_NakamuraCorrespLine(line)
-            break
-        except MatchError:
-            continue
-
-    return nakamuracorrespline
-
-
-class NakamuraCorrespFile(object):
-    """
-    Class for representing nakamura's corresp.txt Files
-    """
-
-    def __init__(self, filename):
-        self.name = filename
-
-        with open(filename) as f:
-            self.lines = np.array(
-                [parse_nakamuracorrespline(line) for line in f.read().splitlines()]
-            )
-
-    @property
-    def note_pairs(self):
-        raise NotImplementedError
-
-    @property
-    def notes(self):
-        raise NotImplementedError
-
-    def iter_notes(self):
-        """
-        Iterate over all note pairs
-        """
-        for x in self.lines:
-            if x is None:
-                continue
-            else:
-                yield x
-
-    @property
-    def snotes(self):
-        raise NotImplementedError
-
-    @property
-    def sustain_pedal(self):
-        raise NotImplementedError
-
-    @property
-    def note_arrays(self):
-        """
-        generate a tuple of (performance-, score-) note arrays with pitch,
-        id and onset information (in seconds).
-
-        Returns
-        -------
-        (array_performance, array_score) : tuple
-            a tuple of structured arrays
-
-        """
-        fields = [("onset_sec", "f4"), ("pitch", "i4"), ("id", "U256")]
-
-        note_array0 = []
-        note_array1 = []
-        for line in self.iter_notes():
-            if line.id0 != "*":
-                note_array0.append((line.onset0, line.pitch0, line.id0))
-            if line.id1 != "*":
-                note_array1.append((line.onset1, line.pitch1, line.id1))
-        ar0 = np.array(note_array0, dtype=fields)
-        ar1 = np.array(note_array1, dtype=fields)
-        return ar0, ar1
-
-    @property
-    def alignment(self):
-        result = []
-        for line in self.iter_notes():
-            if line.id0 == "*":
-                result.append(
-                    dict(label="deletion",
-                         performance_id=line.id0,
-                         score_id=line.id1)
-                )
-            elif line.id1 == "*":
-                result.append(
-                    dict(label="insertion",
-                         performance_id=int(line.id0),
-                         score_id=line.id1)
-                )
-            else:
-                result.append(
-                    dict(label="match",
-                         performance_id=int(line.id0),
-                         score_id=line.id1)
-                )
-        return result
-
-
-def load_nakamuracorresp(fn, pedal_threshold=64, first_note_at_zero=False):
-    """Load a corresp file as returned by Nakamura et al.'s MIDI to MIDI alignment.
+def load_nakamuraspr(fn):
+    """Load a spr file as returned by Nakamura et al.'s alignment methods.
 
     Fields of the file format as specified in [8]_:
-    (ID) (onset time) (spelled pitch) (integer pitch) (onset velocity)
+    ID (onset time) (offset time) (spelled pitch) (onset velocity)
+    (offset velocity) channel
+
+    These files contain extra information not included in match or corresp files,
+    particularly duration and pedal information, and can be used to complement the
+    information from the `load_nakamuracorresp` or `load_nakamuramatch`.
 
     Parameters
     ----------
     fn : str
         The nakamura match.txt-file
-    pedal_threshold : int, optional
-        Threshold for adjusting sound off of the performed notes using
-        pedal information. Defaults to 64.
-    first_note_at_zero : bool, optional
-        When True the note_on and note_off times in the performance
-        are shifted to make the first note_on time equal zero.
 
     Returns
     -------
-    array_performance : structured array
-        structured array of performed notes
-    array_score : structured array
-        structured array of score notes
-    alignment : list
-        The score--performance alignment, a list of dictionaries
-    """
-    cf = NakamuraCorrespFile(fn)
-    array_performance, array_score = cf.note_arrays
-    alignment = cf.alignment
-    return array_performance, array_score, alignment
+    note_array : structured array
+        structured array with note information
+
+    References
+    ----------
+    .. [8] https://midialignment.github.io/MANUAL.pdf
 
 
-def note_name_to_midi_pitch(notename):
+    TODO
+    ----
+    * Import pedal information
     """
-    Utility function to convert the Nakamura pitch spelling to MIDI pitches
-    """
-    pitch = 0
-    keys = [
-        "A",
-        "B",
-        "C",
-        "D",
-        "E",
-        "F",
-        "G",
-        r"\#",
-        "b",
-        "0",
-        "1",
-        "2",
-        "3",
-        "4",
-        "5",
-        "6",
-        "7",
-        "8",
-        "x",
-        "bb",
+    note_array_dtype = [("onset_sec", "f4"),
+                        ("duration_sec", "f4"),
+                        ("pitch", "i4"),
+                        ("velocity", "i4"),
+                        ("channel", "i4"),
+                        ("id", "U256")]
+    dtype = [
+        ("ID", "U256"),
+        ("Ontime", "f"),
+        ("Offtime", "f"),
+        ("Sitch", "U256"),
+        ("Onvel", "i"),
+        ("Offvel", "i"),
+        ("Channel", "i")
     ]
-    values = [
-        21,
-        23,
-        12,
-        14,
-        16,
-        17,
-        19,
-        1,
-        -1,
-        0,
-        12,
-        24,
-        36,
-        48,
-        60,
-        72,
-        84,
-        96,
-        2,
-        -1,
-    ]
-    for k, v in zip(keys, values):
-        if k in notename:
-            pitch += v
-    return pitch
+
+    pattern = r"(\d+)\t(.+)\t(.+)\t(.+)\t(.+)\t(.+)\t(.+)"
+
+    result = np.fromregex(fn, pattern, dtype=dtype)
+    note_array = np.empty(len(result), dtype=note_array_dtype)
+
+    note_array["id"] = result["ID"]
+    note_array["onset_sec"] = result["Ontime"]
+    note_array["duration_sec"] = result["Offtime"] - result["Ontime"]
+    note_array["pitch"] = np.array(
+        [note_name_to_midi_pitch(n.replace('#', r'\#'))
+         for n in result["Sitch"]])
+    note_array["velocity"] = result["Onvel"]
+    note_array["channel"] = result["Channel"]
+
+    return note_array
