+++ tests/test_nakamura.py
@@ -1,40 +1,50 @@
 """
 
-This file contains test functions for the import of Nakamura et al.'s match and corresp file formats.
+This file contains test functions for the import of Nakamura et al.'s match and
+corresp file formats.
 
 """
 
-import logging
 import unittest
-from tempfile import TemporaryFile
-from partitura.io.importnakamura import parse_nakamuracorrespline, load_nakamuracorresp
+from partitura.io.importnakamura import (
+    load_nakamuracorresp,
+    load_nakamuramatch
+)
 
 from . import NAKAMURA_IMPORT_TESTFILES
-
-# from partitura.io.importnakamura import ?
-from partitura import load_nakamuracorresp, load_nakamuramatch
 
 
 class TestLoadNakamura(unittest.TestCase):
     def test_import(self):
         for fn in NAKAMURA_IMPORT_TESTFILES:
             if "corresp" in fn:
-                # read file
-                with open(fn) as f:
-
-                    file_contents = [
-                        (l, parse_nakamuracorrespline(l)) for l in f.read().splitlines()
-                    ]
-
-                for fc in file_contents:
-                    # assert that the lines are correctly encoded
-                    self.assertTrue(fc[0], fc[1].corresp_line)
-
-                # TODO: further tests on load_nakamuracorresp:
-                # corresp = load_nakamuracorresp(fn)
+                performance, score, align = load_nakamuracorresp(fn)
+                performance_ids = set(
+                    item["performance_id"]
+                    for item in align
+                    if item["label"] in ("match", "insertion")
+                )
+                score_ids = set(
+                    item["score_id"]
+                    for item in align
+                    if item["label"] in ("match", "deletion")
+                )
+                self.assertTrue(len(performance_ids) == len(performance))
+                self.assertTrue(len(score_ids) == len(score))
             else:
-                # TODO: test match
-                pass
+                performance, score, align = load_nakamuramatch(fn)
+                performance_ids = set(
+                    item["performance_id"]
+                    for item in align
+                    if item["label"] in ("match", "insertion")
+                )
+                score_ids = set(
+                    item["score_id"]
+                    for item in align
+                    if item["label"] in ("match", "deletion")
+                )
+                self.assertTrue(len(performance_ids) == len(performance))
+                self.assertTrue(len(score_ids) == len(score))
 
 
 if __name__ == "__main__":
